🧪 Test: extract gcal.sync.processor into separate test files
